ipq40xx: add support for USB 3.0 via dwc3
This patch adds code for the two USB-PHYs (baldur and
uniphy). Furthermore, it adds a binding to dwc3-of-simple
which is enough to get the host-mode working.

There isn't any support for the USB 3.0 LOS calibration
(as this requires SCM support for this platform) as well
as the gadget-mode. These will have to be added at a later
time, once a device needs them.

These were taken from Matthew McClintock tree:
<https://source.codeaurora.org/quic/qsdk/mmcclint-qca/>

Signed-off-by: Christian Lamparter <chunkeey@gmail.com>